@@ -727,6 +727,96 @@ const achievements: Achievement[] = [
 		reaction: '108',
 		value: 5,
 	},
+	{
+		id: 'reaction-nameraka_rotating_rolling_on_the_floor_laughing-5',
+		difficulty: 'hard',
+		title: '自身が転んでもおかしいお年頃',
+		condition: '1つのメッセージに対し5人以上から:nameraka_rotating_rolling_on_the_floor_laughing:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'nameraka_rotating_rolling_on_the_floor_laughing',
+		value: 5,
+	},
+	{
+		id: 'reaction-appingorou-5',
+		difficulty: 'hard',
+		title: 'インドのナンごをたべるんご',
+		condition: '1つのメッセージに対し5人以上から:appingorou:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'appingorou',
+		value: 5,
+	},
+	{
+		id: 'reaction-neko-5',
+		difficulty: 'hard',
+		title: 'なんだ猫...か...?',
+		condition: '1つのメッセージに対し5人以上から:neko:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'neko',
+		value: 5,
+	},
+	{
+		id: 'reaction-joyman-takagi-5',
+		difficulty: 'hard',
+		title: '解除それ以上',
+		condition: '1つのメッセージに対し5人以上から:joyman-takagi:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'joyman-takagi',
+		value: 5,
+	},
+	{
+		id: 'reaction-undakeno-5',
+		difficulty: 'hard',
+		title: 'ここにいられる奇跡に感謝',
+		condition: '1つのメッセージに対し5人以上から:undakeno:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'undakeno',
+		value: 5,
+	},
+	{
+		id: 'reaction-mounttoru-5',
+		difficulty: 'hard',
+		title: '実績取るやん',
+		condition: '1つのメッセージに対し5人以上から:mounttoru:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'mounttoru',
+		value: 5,
+	},
+	{
+		id: 'reaction-ichihime-zekkocho-5',
+		difficulty: 'hard',
+		title: 'テンパイ即リー、速攻にゃ速攻ーっ！',
+		condition: '1つのメッセージに対し5人以上から:ichihime-zekkocho:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'ichihime-zekkocho',
+		value: 5,
+	},
+	{
+		id: 'reaction-lifestyle-breaking-5',
+		difficulty: 'hard',
+		title: '起床時間乱数生成器',
+		condition: '1つのメッセージに対し5人以上から:lifestyle-breaking:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'lifestyle-breaking',
+		value: 5,
+	},
+	{
+		id: 'reaction-eminem-5',
+		difficulty: 'hard',
+		title: 'エミネムさんが実績について教えてくれるそうです',
+		condition: '1つのメッセージに対し5人以上から:eminem:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'eminem',
+		value: 5,
+	},
+	{
+		id: 'reaction-tukutte-5',
+		difficulty: 'hard',
+		title: '生みの楽しみ',
+		condition: '1つのメッセージに対し5人以上から:tukutte:リアクションをもらう',
+		category: 'reactions',
+		reaction: 'tukutte',
+		value: 5,
+	},
 
 	// sushi-bot
 
@@ -800,6 +890,79 @@ const achievements: Achievement[] = [
 		condition: '週間起床ランキングで720点以上を獲得する',
 		category: 'sushi-bot',
 	},
+	{
+		id: 'asa-week-0',
+		difficulty: 'hard',
+		title: '生活リズム・完全敗北',
+		condition: '週間起床ランキングで0点以下を獲得し、ランキングに掲載される',
+		category: 'sushi-bot',
+	},
+	{
+		id: 'asa-week-perfect',
+		difficulty: 'professional',
+		title: '生活リズム・オールパーフェクト',
+		condition: '週間起床ランキングで満点を獲得する',
+		category: 'sushi-bot',
+	},
+	{
+		id: 'first-exercise',
+		difficulty: 'easy',
+		title: '一日一善体力づくり',
+		condition: '運動する',
+		category: 'sushi-bot',
+	},
+	{
+		id: 'everyday-exercise-week',
+		difficulty: 'hard',
+		title: 'エクササイズ週間皆勤賞',
+		condition: '週七日運動する',
+		category: 'sushi-bot',
+	},
+	{
+		id: 'exercise-30',
+		difficulty: 'medium',
+		title: '一月分の運動',
+		condition: '30日運動する',
+		category: 'sushi-bot',
+		counter: 'exercise-cumulative',
+		value: 30,
+	},
+	{
+		id: 'exercise-60',
+		difficulty: 'hard',
+		title: '二月分の運動',
+		condition: '60日運動する',
+		category: 'sushi-bot',
+		counter: 'exercise-cumulative',
+		value: 60,
+	},
+	{
+		id: 'exercise-100',
+		difficulty: 'hard',
+		title: 'チリも積もれば筋となる',
+		condition: '100日運動する',
+		category: 'sushi-bot',
+		counter: 'exercise-cumulative',
+		value: 100,
+	},
+	{
+		id: 'exercise-180',
+		difficulty: 'professional',
+		title: 'エクササイズ・ハーフイヤー・セレブレーション',
+		condition: '180日運動する',
+		category: 'sushi-bot',
+		counter: 'exercise-cumulative',
+		value: 180,
+	},
+	{
+		id: 'exercise-365',
+		difficulty: 'professional',
+		title: 'イチロー・スズキ',
+		condition: '365日運動する',
+		category: 'sushi-bot',
+		counter: 'exercise-cumulative',
+		value: 365,
+	},
 
 	// dajare
 
@@ -840,6 +1003,56 @@ const achievements: Achievement[] = [
 		title: 'スーモ',
 		condition: 'pockyから20文字以上の返答を得る',
 		category: 'pocky',
+	},
+	{
+		id: 'pocky-on-nov-11',
+		difficulty: 'medium',
+		title: '最後までチョコたっぷり',
+		condition: '11月11日にpockyに反応される',
+		category: 'pocky',
+	},
+	{
+		id: 'pockygame-on-nov-11',
+		difficulty: 'medium',
+		title: 'Stick to fun!',
+		condition: '11月11日にポッキーゲームに勝利する',
+		category: 'pocky',
+	},
+	{
+		id: 'pockygame-win-1',
+		difficulty: 'easy',
+		title: '堅焼きプレッツェル',
+		condition: 'ポッキーゲームに1回勝利する',
+		category: 'pocky',
+		counter: 'pockygame-win',
+		value: 1,
+	},
+	{
+		id: 'pockygame-win-10',
+		difficulty: 'medium',
+		title: 'ラッキースティック',
+		condition: 'ポッキーゲームに10回勝利する',
+		category: 'pocky',
+		counter: 'pockygame-win',
+		value: 10,
+	},
+	{
+		id: 'pockygame-win-100',
+		difficulty: 'hard',
+		title: 'フランダブルショコラ',
+		condition: 'ポッキーゲームに100回勝利する',
+		category: 'pocky',
+		counter: 'pockygame-win',
+		value: 100,
+	},
+	{
+		id: 'pockygame-win-1000',
+		difficulty: 'professional',
+		title: 'MIKADO',
+		condition: 'ポッキーゲームに1000回勝利する',
+		category: 'pocky',
+		counter: 'pockygame-win',
+		value: 1000,
 	},
 
 	// tahoiya
@@ -1375,6 +1588,51 @@ const achievements: Achievement[] = [
 		category: 'ricochet-robots',
 	},
 
+	// hitandblow
+
+	{
+		id: 'hitandblow-play',
+		difficulty: 'easy',
+		title: '0 Hit 1 Blow',
+		condition: 'Hit & Blow で遊ぶ',
+		category: 'hitandblow',
+	},
+	{
+		id: 'hitandblow-clear',
+		difficulty: 'easy',
+		title: '1 Hit 0 Blow',
+		condition: 'Hit & Blow をクリアする',
+		category: 'hitandblow',
+	},
+	{
+		id: 'hitandblow-clear-6digits-or-more',
+		difficulty: 'medium',
+		title: '6 Hit 0 Blow',
+		condition: 'Hit & Blow をクリアする (6桁以上)',
+		category: 'hitandblow',
+	},
+	{
+		id: 'hitandblow-clear-10digits',
+		difficulty: 'hard',
+		title: '10 Hit 0 Blow',
+		condition: 'Hit & Blow をクリアする (10桁)',
+		category: 'hitandblow',
+	},
+	{
+		id: 'hitandblow-clear-once-3digits-or-more',
+		difficulty: 'professional',
+		title: 'guess問が得意',
+		condition: 'Hit & Blow を一回のコールでクリアする (3桁以上)',
+		category: 'hitandblow',
+	},
+	{
+		id: 'hitandblow-clear-10digits-within-5min',
+		difficulty: 'professional',
+		title: '実質ソルバ',
+		condition: 'Hit & Blow (10桁) を5分以内にクリアする',
+		category: 'hitandblow',
+	},
+
 	// ahokusa
 
 	{
@@ -1628,6 +1886,37 @@ const achievements: Achievement[] = [
 		value: 20,
 	},
 
+	// pwnyaa
+
+	{
+		id: 'pwnyaa-tw-half',
+		difficulty: 'easy',
+		title: 'ポーンではなくパウンですよ!',
+		condition: 'pwnable.twで半分以上の問題を解く',
+		category: 'pwnyaa',
+	},
+	{
+		id: 'pwnyaa-tw-complete',
+		difficulty: 'medium',
+		title: 'ぱあんぱいんぱうんぱえんぱおん',
+		condition: 'pwnable.twを全完する',
+		category: 'pwnyaa',
+	},
+	{
+		id: 'pwnyaa-xyz-half',
+		difficulty: 'baby',
+		title: '始めることが、大事です',
+		condition: 'pwnable.xyzで半分以上の問題を解く',
+		category: 'pwnyaa',
+	},
+	{
+		id: 'pwnyaa-xyz-complete',
+		difficulty: 'baby',
+		title: '続けることが、大事です',
+		condition: 'pwnable.xyzを全完する',
+		category: 'pwnyaa',
+	},
+
 	// anime
 
 	{
@@ -1841,6 +2130,15 @@ const achievements: Achievement[] = [
 		value: 150,
 	},
 	{
+		id: 'crossword-wins-1000',
+		difficulty: 'professional',
+		title: 'サザンクロス',
+		condition: 'crosswordで1マス以上のマスを開け、盤面を1000回クリアする',
+		category: 'crossword',
+		counter: 'crossword-wins',
+		value: 1000,
+	},
+	{
 		id: 'crossword-solo',
 		difficulty: 'hard',
 		title: 'クロスワード・完投勝利',
@@ -1868,6 +2166,63 @@ const achievements: Achievement[] = [
 		condition: 'crosswordで制限時間が切れる2秒以内に、クリアを決定する最後の単語を回答する',
 		category: 'crossword',
 	},
+	{
+		id: 'crossword-game-time-le-quarter',
+		difficulty: 'hard',
+		title: 'クロスワード・ゲームセッター',
+		condition: 'crosswordで制限時間の1/4以下の時間でクリアしたゲームにおいて、1マス以上のマスを開ける',
+		category: 'crossword',
+	},
+	{
+		id: 'crossword-misses-ge-20',
+		difficulty: 'medium',
+		title: 'クロスワード・フリースインガー',
+		condition: 'crosswordでクリアしたゲームにおいて、20個以上の誤答を行い、一度も正解しない',
+		category: 'crossword',
+	},
+	{
+		id: 'crossword-contributors-ge-11',
+		difficulty: 'professional',
+		title: 'クロスワード・オールスターゲーム',
+		condition: 'crosswordで1ゲームに11人以上が正答した状態でクリアしたゲームにおいて、1マス以上のマスを開ける',
+		category: 'crossword',
+	},
+	{
+		id: 'grossword-wins-1',
+		difficulty: 'hard',
+		title: 'シングル・グロス',
+		condition: 'grosswordで1マス以上のマスを開け、盤面をクリアする',
+		category: 'crossword',
+		counter: 'grossword-wins',
+		value: 1,
+	},
+	{
+		id: 'grossword-wins-12',
+		difficulty: 'hard',
+		title: 'ダズン・グロス',
+		condition: 'grosswordで1マス以上のマスを開け、盤面を12回クリアする',
+		category: 'crossword',
+		counter: 'grossword-wins',
+		value: 12,
+	},
+	{
+		id: 'grossword-wins-144',
+		difficulty: 'professional',
+		title: 'グロス・グロス',
+		condition: 'grosswordで1マス以上のマスを開け、盤面を144回クリアする',
+		category: 'crossword',
+		counter: 'grossword-wins',
+		value: 144,
+	},
+	{
+		id: 'grossword-wins-1728',
+		difficulty: 'professional',
+		title: 'グレート・グロス・グロス',
+		condition: 'grosswordで1マス以上のマスを開け、盤面を1728回クリアする',
+		category: 'crossword',
+		counter: 'grossword-wins',
+		value: 1728,
+	},
 
 	// denim (bcr)
 	{
@@ -1875,16 +2230,15 @@ const achievements: Achievement[] = [
 		difficulty: 'medium',
 		title: 'ガチ割れ行くぜ！',
 		condition: '白鍵を4567鍵に寄せ、デニム譜面を割る',
-		category: 'bcr'
+		category: 'bcr',
 	},
 	{
 		id: 'bcr-denim-reversplit',
 		difficulty: 'medium',
 		title: '逆割れ行くぜ！',
 		condition: '白鍵を1234鍵に寄せ、デニム譜面を2P割れにする',
-		category: 'bcr'
+		category: 'bcr',
 	},
-
 
 	// achievements
 
